splitting helper into bin-helper & lib-helper
