+++ lib/create-file.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const templateFactory = require('./template-factory');
-const { writeFile } = require('../bin/helpers');
+const { writeFile } = require('../lib/lib-helpers');
 const path = require('path');
 
 const createComponent = (name, isDumb, createDir, overwriteFile, options) => {
