+++ lib/lib-helpers.js
@@ -1,22 +1,8 @@
 'use strict';
 
-const Configstore = require('configstore');
-const conf = new Configstore('napp-config');
 const fs = require('fs-extra');
 
 const helper = {
-    changeAllSettings(bool, store = conf) {
-        store.set({
-            componentWillMount: bool,
-            componentWillReceiveProps: bool,
-            shouldComponentUpdate: bool,
-            componentWillUpdate: bool,
-            componentDidMount: bool,
-            componentDidUpdate: bool,
-            componentWillUnmount: bool,
-            componentDidCatch: bool
-        });
-    },
     writeFile(file, template, overwriteFile, createDir=null) {
         fs.pathExists(file, (err, exists) => {
             if (err) {
